Clean up theme json script
