+++ src/components/ids-theme-switcher/demos/token-table.html
@@ -22,8 +22,8 @@
       <ids-layout-flex-item grow="1" overflow="hidden">
         <ids-data-grid 
           id="data-theme" 
-          suppress-tooltips="true" 
-          row-selection="multiple" 
+          suppress-tooltips="true"
+          row-selection="single" 
           tree-grid="true" 
           auto-fit="true" 
           virtual-scroll="true"
